requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
duckdb==0.9.2
pandas==2.1.4
numpy==1.25.2
transformers==4.36.2
torch==2.1.1
scikit-learn==1.3.2
snscrape==0.7.0.20230622
praw==7.7.1
PyGithub==2.1.1
telethon==1.34.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0
aiohttp==3.9.1
matplotlib==3.8.2
joblib==1.3.2

